apps/backend/pyproject.toml
[tool.poetry]
name = "cherrypick-backend"
version = "0.1.0"
description = "Cherrypick Backend - Intelligent Resume Assembly Engine"
authors = []
readme = "README.md"
packages = [{include = "cherrypick_backend"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.0"}
asyncpg = "^0.30.0"
pydantic-settings = "^2.6.0"
python-dotenv = "^1.0.0"
chromadb = "^1.4.0"
httpx = "^0.28.1"
python-dateutil = "^2.8.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
black = "^24.10.0"
ruff = "^0.8.0"
mypy = "^1.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
